Refactored reference to patient id to identityId in a way that can be undo easily

+++ app/src/main/java/com/example/a101_monitoring/ui/RegisterPatientFragment.kt
@@ -124,7 +124,7 @@ class RegisterPatientFragment : Fragment() {
 
     fun onRegisterClicked(view: View) {
         registerPatientViewModel.registerPatient(
-            registered_patient_id.text.toString().toInt(),
+            registered_patient_id.text.toString(),
             1,
             "1",
             "1",
